feat: Really basic json-rpc server impl
Requires the `AUTH_TOKEN` env var to be set. Currently only supports a
single rpc:

```json
{
  "rpc": "miner-info"
}
```

If we want to keep building in this way and build out more underneath
it, like adding database aggregation with some workers, could be the way
to go. Otherwise, can scrap this for another approach / doing more logic
client-side.